Extract package and fix typo (fixes #1)
- extracted pkg so that the code can be used as a library
- corrected typo in README
- corrected Golang comments - comments on structs should start
with the name of the struct (VSCode shows these hints)

Tested locally on MacOS
